Baseline Notebook + CSV Files
James Ze Wang[1] (Chinese: 王则; born 1972) is a Chinese-American computer scientist. He is a distinguished professor of the College of Information Sciences and Technology at Pennsylvania State University.[2] He is also an affiliated professor of the Molecular, Cellular, and Integrative Biosciences Program; the Computational Science Graduate Minor; and the Social Data Analytics Graduate Program.[2] He is co-director of the Intelligent Information Systems Laboratory.[2][3] He was a visiting professor of the Robotics Institute at Carnegie Mellon University from 2007 to 2008.[3] In 2011 and 2012, he served as a program manager in the Office of International Science and Engineering at the National Science Foundation. He is the second son of Chinese mathematician Wang Yuan.